feat: implement sound alerts for system notifications and fix multiple POS issues
- Added pos_notification_sound_id to system_settings configuration
- Implemented audio playback in NotificationContainer reactive to localStorage
- Updated ConfigSoundsCard with POS sound selector and immediate sync
- Fixed cash drawer IPC handler in Electron
- Corrected email recipient in ShiftService
- Fixed decimal precision in inventory calculations

+++ components/admin/inventarios/InventariosLayout.tsx
@@ -605,7 +605,7 @@ export const InventariosLayout: React.FC<InventariosLayoutProps> = ({ initialTab
                     unit_measure: newProduct.unit_measure || '',
                     presentation_unit: newProduct.presentation_unit || '',
                     // Sincronizar SIEMPRE ambos campos con el valor que el usuario ingresó en pantalla
-                    portions: parseInt(newProduct.conversion_factor) || parseInt(newProduct.portions) || 1,
+                    portions: parseFloat(newProduct.conversion_factor) || parseFloat(newProduct.portions) || 1,
                     conversion_factor: parseFloat(newProduct.conversion_factor) || parseFloat(newProduct.portions) || 1,
                     cost_price: parseFloat(newProduct.cost_price) || 0,
                     price: 0,
